refactor: add proper __init__.py and update imports
